requirements.txt
requests==2.31.0
python-dotenv==1.0.0
rich==13.7.0
PyYAML==6.0.1
streamlit==1.29.0
pandas==2.1.4
orjson==3.9.10
feedparser==6.0.10
python-dateutil==2.8.2
